backend/requirements.txt
# Common requirements for all backend services
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Azure SDKs
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-servicebus==7.11.4
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-ai-inference>=1.0.0b9

# Database
pyodbc==5.0.1
sqlalchemy==2.0.23

# AI/ML
openai>=1.3.8
semantic-kernel>=0.9.6b1

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
httpx==0.25.2